--- app/src/main/java/pw/prsk/goodfood/utils/IngredientDiffUtilCallback.kt
@@ -1,21 +1,0 @@
-package pw.prsk.goodfood.utils
-
-import androidx.recyclerview.widget.DiffUtil
-import pw.prsk.goodfood.data.IngredientWithMeta
-
-class IngredientDiffUtilCallback(
-    private val oldList: List<IngredientWithMeta>,
-    private val newList: List<IngredientWithMeta>
-) : DiffUtil.Callback() {
-    override fun getOldListSize(): Int = oldList.size
-
-    override fun getNewListSize(): Int = newList.size
-
-    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return oldList[oldItemPosition].product.id == newList[newItemPosition].product.id
-    }
-
-    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return oldList[oldItemPosition] == newList[newItemPosition]
-    }
-}
