--- app/src/main/java/pw/prsk/goodfood/utils/MealDiffUtilCallback.kt
@@ -1,19 +1,0 @@
-package pw.prsk.goodfood.utils
-
-import androidx.recyclerview.widget.DiffUtil
-import pw.prsk.goodfood.data.Meal
-
-class MealDiffUtilCallback(private val oldList: List<Meal>, private val newList: List<Meal>) :
-    DiffUtil.Callback() {
-    override fun getOldListSize(): Int = oldList.size
-
-    override fun getNewListSize(): Int = newList.size
-
-    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return oldList[oldItemPosition].id == newList[newItemPosition].id
-    }
-
-    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return oldList[oldItemPosition] == newList[newItemPosition]
-    }
-}
