--- app/src/main/java/pw/prsk/goodfood/utils/ProductDiffUtilCallback.kt
@@ -1,19 +1,0 @@
-package pw.prsk.goodfood.utils
-
-import androidx.recyclerview.widget.DiffUtil
-import pw.prsk.goodfood.data.Product
-
-class ProductDiffUtilCallback(private val oldList: List<Product>, private val newList: List<Product>) :
-    DiffUtil.Callback() {
-    override fun getOldListSize(): Int = oldList.size
-
-    override fun getNewListSize(): Int = newList.size
-
-    override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return oldList[oldItemPosition].id == newList[newItemPosition].id
-    }
-
-    override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
-        return oldList[oldItemPosition] == newList[newItemPosition]
-    }
-}
