+++ app/src/main/res/layout/item_ingredient.xml
@@ -5,8 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="horizontal"
-    android:paddingTop="9dp"
-    android:paddingBottom="9dp" >
+    android:paddingVertical="9dp" >
 
     <TextView
         android:layout_width="wrap_content"
